development/python/database/requirements.txt
psycopg2-binary
mysql-connector-python
snowflake-connector-python
singlestoredb